Remove logs. Refactor Components to small components.
