+++ src/components/chart/utils/ChartLabel.js
@@ -1,0 +1,4 @@
+//REMIND 오늘 기준으로 라벨 업데이트 하는 로직 구현
+export const monthlyLabel = ['January', 'February', 'March', 'April', 'May', 'June', 'July'];
+
+export const weeklyLabel = ['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday', 'Saturday', 'Sunday'];
